style: renaming directory architecture
